requirements.txt
flask==3.0.0
werkzeug==2.3.7
flask-sqlalchemy==3.1.1
mysql-connector-python==8.1.0
flask-login==0.6.3
cryptography==41.0.3
fpdf2==2.7.4
python-dotenv==1.0.0
